pyproject.toml
[project]
name = "memoryllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "gradio>=6.3.0",
    "plotly>=6.5.2",
    "sentence-transformers>=5.2.0",
    "umap-learn>=0.5.11",
]
